requirements.txt
six==1.15.0
certifi==2020.11.8
urllib3>=1.26.5
python-telegram-bot
python_dateutil==2.8.2
simplejson==3.17.5
